Removed some more STL references.

+++ UserInput/Implementation/KeyBinding.cpp
@@ -1,6 +1,6 @@
 #include <UserInput/KeyBinding.hpp>
 
-KeyBinding::KeyBinding(const uint16_t& newKeyCode, const KeyState& newKeyState) {
+KeyBinding::KeyBinding(const size_t& newKeyCode, const KeyState& newKeyState) {
     keyCode = newKeyCode;
     keyState = newKeyState;
 }
