+++ UserInput/Interface/UserInput/KeyBinding.hpp
@@ -1,14 +1,12 @@
 #pragma once
 
-#include <cstdint>
-
 #include "KeyState.hpp"
 
 struct KeyBinding {
 public:
-    KeyBinding(const uint16_t& newKeyCode, const KeyState& newKeyState);
+    KeyBinding(const size_t& newKeyCode, const KeyState& newKeyState);
 
 public:
-    uint16_t keyCode;
+    size_t keyCode;
     KeyState keyState;
 };
